requirements.txt
Pillow
qrcode
pyttsx3